wrapper added

+++ flaskr/flight/routes.py
@@ -6,8 +6,9 @@ from datetime import datetime
 
 from flaskr.seat.routes import create_seats, delete_seats
 
+from flaskr.auth.routes import login_required, admin_required, super_admin_required
 
-from flaskr.models import Flight, db, City
+from flaskr.models import Flight, db, City, AirLine
 
 from sqlalchemy import select, insert, delete
 
@@ -15,26 +16,30 @@ from sqlalchemy import select, insert, delete
 flight_bp = Blueprint('flights', __name__, url_prefix='/flights')
 
 
+# @flight_bp.get('/')
+# def get_all_flights():
+#     error = None
+
+#     try:
+#         flights = db.session.execute(select(Flight))
+#     except Exception as e:
+#         print(e)
+#         abort(500)
+
+#     return render_template('/index.html', flights=flights)
+
+
 @flight_bp.get('/')
-def get_all_flights():
+def get_flights_by_airline_id():
     error = None
-
+    airline_id = request.args.get('airline_id')
     try:
-        flights = db.session.execute(select(Flight))
-    except Exception as e:
-        print(e)
-        abort(500)
-
-    return render_template('/index.html', flights=flights)
-
-
-@flight_bp.get('/<int:airline_id>/')
-def get_flights_by_airline_id(airline_id):
-    error = None
-
-    try:
-        flights = db.session.execute(
-            select(Flight).filter_by(airline_id=airline_id))
+        if airline_id:
+            flights = db.session.execute(
+                select(Flight).filter_by(airline_id=airline_id)).scalars().all()
+        else:
+            flights = db.session.execute(select(Flight)).scalars().all()
+        print(flights)
     except Exception as e:
         print(e)
         abort(500)
@@ -61,8 +66,12 @@ def search_flights():
 
 
 @flight_bp.post('/new')
+@admin_required
 def create_flight():
     data = request.form.to_dict()
+    data['airline_id'] = db.first_or_404(
+        select(AirLine.id).filter(AirLine.admin_id == g.user.id))
+    print(data['airline_id'])
     data['available_seats'] = data['total_seats']
     data['departure_time'] = datetime(2024, 12, 10, 5, 30)
     error = None
@@ -81,11 +90,16 @@ def create_flight():
 
 
 @flight_bp.route('/<int:id>/delete', methods=['POST'])
+@admin_required
 def delete_flight(id):
     error = None
 
     try:
+        airline_id = db.first_or_404(
+            select(AirLine.id).filter(AirLine.admin_id == g.user.id))
         flight = db.get_or_404(Flight, id)
+        if (flight.airline_id != airline_id):
+            abort(401)
         db.session.delete(flight)
         delete_seats(id)
         db.session.commit()
